// client-angular/src/app/shared/Users.interface.ts
import { FormControl } from "@angular/forms";

export interface Users {
  name: FormControl<string>;
  username: FormControl<string>;
  age: FormControl<string>;
  nationality: FormControl<string>;
}

export interface Geo {
    lat: string;
    lng: string;
  }
  
  export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
  }
  
  export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
  }
  
  export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: Address;
    phone: string;
    website: string;
    company: Company;
  }
  
  export interface Post {
    userId: number;
    id: number;
    title: string;
    body: string;
  }
  
  export interface UserWithPosts extends User {
    posts: Post[];
  }